requirements.txt
altair==6.0.0
annotated-types==0.7.0
anthropic==0.92.0
anyio==4.13.0
attrs==26.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
bracex==2.6
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
deepagents==0.5.1
distro==1.9.0
docstring_parser==0.17.0
dotenv==0.9.9
filetype==1.2.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.1
google-genai==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.15
langchain-anthropic==1.4.0
langchain-core==1.2.28
langchain-google-genai==4.2.1
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.7.27
markdownify==1.2.2
MarkupSafe==3.0.3
narwhals==2.19.0
numpy==2.4.4
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
pandas==3.0.2
pillow==12.2.0
protobuf==7.34.1
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyspark==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
streamlit==1.56.0
tavily-python==0.7.23
tenacity==9.1.4
tiktoken==0.12.0
toml==0.10.2
tornado==6.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uuid_utils==0.14.1
watchdog==6.0.0
wcmatch==10.1
websockets==16.0
xxhash==3.6.0
zstandard==0.25.0
